refactor(phase-19)(pr-4a): social/ws/editor service 파일 분할
F-go-3 해소 — 3 인프라 파일 500줄 한도 초과 해소.

**Before → After**
- domain/social/service.go: 759 → 43 (friends/blocks/chat_rooms/chat_messages.go로 분리)
- ws/hub.go: 649 → 175 (hub_lifecycle.go + hub_broadcast.go 분리)
- domain/editor/service.go: 505 → 185 (themes.go + publication.go + helpers.go 분리)

Package 유지. 공개 API 시그니처 불변. 전 패키지 go test ok, vet clean.
