+++ apps/server/internal/domain/editor/service.go
@@ -1,25 +1,30 @@
+// Package editor implements the in-game editor domain (themes, characters,
+// maps, locations, clues, and more). The domain is deliberately split
+// across sibling files:
+//
+//   - service.go            interface + constructor + Request/Response DTOs
+//   - themes.go             Theme CRUD (Create/Update/Delete/Get/List)
+//   - publication.go        Publish / Unpublish / SubmitForReview
+//   - helpers.go            getOwnedTheme, pgtype converters, slug/unique-violation helpers, toThemeResponse
+//   - service_character.go  character-scoped methods
+//   - service_location.go   map/location-scoped methods
+//   - service_clue.go       clue-scoped methods
+//   - service_config.go     module config + content methods
+//   - service_validation.go theme validation
+//   - clue_edge_service.go  Phase 20 PR-4 unified clue edge schema
+//
+// Callers depend only on the public Service interface defined here.
 package editor
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/hex"
 	"encoding/json"
-	"errors"
-	"fmt"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgconn"
-	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/jackc/pgx/v5/pgxpool"
-
 	"github.com/rs/zerolog"
 
-	"github.com/mmp-platform/server/internal/apperror"
 	"github.com/mmp-platform/server/internal/db"
 )
 
@@ -177,329 +182,4 @@ func NewService(q *db.Queries, pool *pgxpool.Pool, logger zerolog.Logger) Servic
 		pool:   pool,
 		logger: logger.With().Str("domain", "editor").Logger(),
 	}
-}
-
-// --- Theme ops ---
-
-func (s *service) CreateTheme(ctx context.Context, creatorID uuid.UUID, req CreateThemeRequest) (*ThemeResponse, error) {
-	var theme db.Theme
-	var err error
-
-	for attempt := 0; attempt < 3; attempt++ {
-		slug := generateSlug(req.Title)
-		theme, err = s.q.CreateTheme(ctx, db.CreateThemeParams{
-			CreatorID:   creatorID,
-			Title:       req.Title,
-			Slug:        slug,
-			Description: ptrToText(req.Description),
-			CoverImage:  ptrToText(req.CoverImage),
-			MinPlayers:  req.MinPlayers,
-			MaxPlayers:  req.MaxPlayers,
-			DurationMin: req.DurationMin,
-			Price:       req.Price,
-			CoinPrice:   req.CoinPrice,
-			ConfigJson:  json.RawMessage("{}"),
-		})
-		if err == nil {
-			return toThemeResponse(theme), nil
-		}
-		if isUniqueViolation(err) {
-			continue
-		}
-		break
-	}
-	s.logger.Error().Err(err).Msg("failed to create theme")
-	return nil, apperror.Internal("failed to create theme")
-}
-
-func (s *service) UpdateTheme(ctx context.Context, creatorID, themeID uuid.UUID, req UpdateThemeRequest) (*ThemeResponse, error) {
-	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
-	if err != nil {
-		return nil, err
-	}
-
-	var updated db.Theme
-	for attempt := 0; attempt < 3; attempt++ {
-		slug := generateSlug(req.Title)
-		updated, err = s.q.UpdateTheme(ctx, db.UpdateThemeParams{
-			ID:          theme.ID,
-			Title:       req.Title,
-			Slug:        slug,
-			Description: ptrToText(req.Description),
-			CoverImage:  ptrToText(req.CoverImage),
-			MinPlayers:  req.MinPlayers,
-			MaxPlayers:  req.MaxPlayers,
-			DurationMin: req.DurationMin,
-			Price:       req.Price,
-			CoinPrice:   req.CoinPrice,
-			Version:     theme.Version,
-		})
-		if err == nil {
-			return toThemeResponse(updated), nil
-		}
-		if isUniqueViolation(err) {
-			continue
-		}
-		break
-	}
-	if errors.Is(err, pgx.ErrNoRows) {
-		return nil, apperror.Conflict("theme was modified by another session")
-	}
-	s.logger.Error().Err(err).Msg("failed to update theme")
-	return nil, apperror.Internal("failed to update theme")
-}
-
-func (s *service) DeleteTheme(ctx context.Context, creatorID, themeID uuid.UUID) error {
-	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
-	if err != nil {
-		return err
-	}
-	if theme.Status != "DRAFT" {
-		return apperror.BadRequest("only draft themes can be deleted")
-	}
-	if err := s.q.DeleteTheme(ctx, theme.ID); err != nil {
-		s.logger.Error().Err(err).Msg("failed to delete theme")
-		return apperror.Internal("failed to delete theme")
-	}
-	return nil
-}
-
-func (s *service) ListMyThemes(ctx context.Context, creatorID uuid.UUID) ([]ThemeSummary, error) {
-	themes, err := s.q.ListThemesByCreator(ctx, creatorID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to list themes")
-		return nil, apperror.Internal("failed to list themes")
-	}
-	out := make([]ThemeSummary, len(themes))
-	for i, t := range themes {
-		out[i] = ThemeSummary{
-			ID:         t.ID,
-			Title:      t.Title,
-			Status:     t.Status,
-			MinPlayers: t.MinPlayers,
-			MaxPlayers: t.MaxPlayers,
-			CoinPrice:  t.CoinPrice,
-			Version:    t.Version,
-			CreatedAt:  t.CreatedAt,
-		}
-	}
-	return out, nil
-}
-
-func (s *service) GetTheme(ctx context.Context, creatorID, themeID uuid.UUID) (*ThemeResponse, error) {
-	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
-	if err != nil {
-		return nil, err
-	}
-	return toThemeResponse(theme), nil
-}
-
-// PublishTheme directly publishes a theme (admin-only use).
-func (s *service) PublishTheme(ctx context.Context, creatorID, themeID uuid.UUID) (*ThemeResponse, error) {
-	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
-	if err != nil {
-		return nil, err
-	}
-	if theme.Status != "DRAFT" {
-		return nil, apperror.BadRequest("only draft themes can be published")
-	}
-
-	charCount, err := s.q.CountThemeCharacters(ctx, theme.ID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count characters")
-		return nil, apperror.Internal("failed to count characters")
-	}
-	if charCount < int64(theme.MinPlayers) {
-		return nil, apperror.BadRequest(fmt.Sprintf("theme requires at least %d characters, has %d", theme.MinPlayers, charCount))
-	}
-
-	if len(theme.ConfigJson) == 0 || string(theme.ConfigJson) == "{}" || string(theme.ConfigJson) == "null" {
-		return nil, apperror.BadRequest("theme config_json must be set before publishing")
-	}
-
-	updated, err := s.q.UpdateThemeStatus(ctx, db.UpdateThemeStatusParams{
-		ID:     theme.ID,
-		Status: "PUBLISHED",
-	})
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to publish theme")
-		return nil, apperror.Internal("failed to publish theme")
-	}
-	return toThemeResponse(updated), nil
-}
-
-// UnpublishTheme moves a published theme to UNPUBLISHED status.
-func (s *service) UnpublishTheme(ctx context.Context, creatorID, themeID uuid.UUID) (*ThemeResponse, error) {
-	theme, err := s.getOwnedTheme(ctx, creatorID, themeID)
-	if err != nil {
-		return nil, err
-	}
-	if theme.Status != "PUBLISHED" {
-		return nil, apperror.BadRequest("only published themes can be unpublished")
-	}
-
-	updated, err := s.q.UnpublishTheme(ctx, theme.ID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to unpublish theme")
-		return nil, apperror.Internal("failed to unpublish theme")
-	}
-	return toThemeResponse(updated), nil
-}
-
-// SubmitForReview submits a theme for admin review.
-// Trusted creators are auto-approved and published immediately.
-func (s *service) SubmitForReview(ctx context.Context, userID, themeID uuid.UUID) (*ThemeResponse, error) {
-	theme, err := s.getOwnedTheme(ctx, userID, themeID)
-	if err != nil {
-		return nil, err
-	}
-
-	if theme.Status != "DRAFT" && theme.Status != "REJECTED" {
-		return nil, apperror.BadRequest("only draft or rejected themes can be submitted for review")
-	}
-
-	charCount, err := s.q.CountThemeCharacters(ctx, theme.ID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to count characters")
-		return nil, apperror.Internal("failed to count characters")
-	}
-	if charCount < int64(theme.MinPlayers) {
-		return nil, apperror.BadRequest(fmt.Sprintf("theme requires at least %d characters, has %d", theme.MinPlayers, charCount))
-	}
-	if len(theme.ConfigJson) == 0 || string(theme.ConfigJson) == "{}" || string(theme.ConfigJson) == "null" {
-		return nil, apperror.BadRequest("theme config_json must be set before submitting")
-	}
-
-	trusted, err := s.q.GetUserTrustedCreator(ctx, userID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to check trusted creator status")
-		return nil, apperror.Internal("failed to check trusted creator status")
-	}
-
-	if trusted {
-		updated, err := s.q.ApproveTheme(ctx, db.ApproveThemeParams{
-			ID:         theme.ID,
-			ReviewedBy: pgtype.UUID{Bytes: userID, Valid: true},
-			ReviewNote: pgtype.Text{String: "auto-approved (trusted creator)", Valid: true},
-		})
-		if err != nil {
-			s.logger.Error().Err(err).Msg("failed to auto-approve theme")
-			return nil, apperror.Internal("failed to auto-approve theme")
-		}
-		return toThemeResponse(updated), nil
-	}
-
-	updated, err := s.q.SubmitThemeForReview(ctx, theme.ID)
-	if err != nil {
-		s.logger.Error().Err(err).Msg("failed to submit theme for review")
-		return nil, apperror.Internal("failed to submit theme for review")
-	}
-	return toThemeResponse(updated), nil
-}
-
-// --- Shared helpers ---
-
-// getOwnedTheme fetches a theme and verifies creator ownership.
-func (s *service) getOwnedTheme(ctx context.Context, creatorID, themeID uuid.UUID) (db.Theme, error) {
-	theme, err := s.q.GetTheme(ctx, themeID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return db.Theme{}, apperror.NotFound("theme not found")
-		}
-		s.logger.Error().Err(err).Msg("failed to get theme")
-		return db.Theme{}, apperror.Internal("failed to get theme")
-	}
-	if theme.CreatorID != creatorID {
-		return db.Theme{}, apperror.Forbidden("you do not own this theme")
-	}
-	return theme, nil
-}
-
-func textToPtr(t pgtype.Text) *string {
-	if !t.Valid {
-		return nil
-	}
-	return &t.String
-}
-
-func ptrToText(s *string) pgtype.Text {
-	if s == nil {
-		return pgtype.Text{}
-	}
-	return pgtype.Text{String: *s, Valid: true}
-}
-
-func int32PtrToPgtype(i *int32) pgtype.Int4 {
-	if i == nil {
-		return pgtype.Int4{}
-	}
-	return pgtype.Int4{Int32: *i, Valid: true}
-}
-
-func pgtypeInt4ToPtr(i pgtype.Int4) *int32 {
-	if !i.Valid {
-		return nil
-	}
-	v := i.Int32
-	return &v
-}
-
-var slugCleanRe = regexp.MustCompile(`[^a-z0-9-]+`)
-
-func generateSlug(title string) string {
-	s := strings.ToLower(strings.TrimSpace(title))
-	s = strings.ReplaceAll(s, " ", "-")
-	s = slugCleanRe.ReplaceAllString(s, "")
-	s = strings.Trim(s, "-")
-	if s == "" {
-		s = "theme"
-	}
-
-	b := make([]byte, 4)
-	if _, err := rand.Read(b); err != nil {
-		// fallback: use timestamp-based suffix
-		return fmt.Sprintf("%s-%d", s, time.Now().UnixMilli()%10000)
-	}
-	return fmt.Sprintf("%s-%s", s, hex.EncodeToString(b)[:4])
-}
-
-// isUniqueViolation returns true if err is a PostgreSQL unique constraint violation (code 23505).
-func isUniqueViolation(err error) bool {
-	var pgErr *pgconn.PgError
-	return errors.As(err, &pgErr) && pgErr.Code == "23505"
-}
-
-func toThemeResponse(t db.Theme) *ThemeResponse {
-	resp := &ThemeResponse{
-		ID:          t.ID,
-		Title:       t.Title,
-		Slug:        t.Slug,
-		Description: textToPtr(t.Description),
-		CoverImage:  textToPtr(t.CoverImage),
-		MinPlayers:  t.MinPlayers,
-		MaxPlayers:  t.MaxPlayers,
-		DurationMin: t.DurationMin,
-		Price:       t.Price,
-		CoinPrice:   t.CoinPrice,
-		Status:      t.Status,
-		Version:     t.Version,
-		CreatedAt:   t.CreatedAt,
-	}
-	if len(t.ConfigJson) > 0 && string(t.ConfigJson) != "null" {
-		resp.ConfigJson = t.ConfigJson
-	}
-	if t.ReviewNote.Valid {
-		s := t.ReviewNote.String
-		resp.ReviewNote = &s
-	}
-	if t.ReviewedAt.Valid {
-		ts := t.ReviewedAt.Time
-		resp.ReviewedAt = &ts
-	}
-	if t.ReviewedBy.Valid {
-		id := t.ReviewedBy.Bytes
-		uid := uuid.UUID(id)
-		resp.ReviewedBy = &uid
-	}
-	return resp
 }
